fix(index): harden direct ADR renderer lifecycle
