+++ scripts/verify/render-index-storage-adr.test.mjs
@@ -2,7 +2,15 @@
 
 import { createHash } from 'node:crypto';
 import { spawnSync } from 'node:child_process';
-import { mkdtempSync, mkdirSync, readFileSync, rmSync, writeFileSync } from 'node:fs';
+import {
+  existsSync,
+  mkdtempSync,
+  mkdirSync,
+  readdirSync,
+  readFileSync,
+  rmSync,
+  writeFileSync,
+} from 'node:fs';
 import { tmpdir } from 'node:os';
 import path from 'node:path';
 import assert from 'node:assert/strict';
@@ -37,7 +45,9 @@ const decisionFlags = {
 
 const readPrototype = (prototype, scaleFactor) => ({
   prototype,
-  schema: prototype === 'typed_eav' ? 'idx_bench_eav' : `idx_bench_${prototype === 'hot_projection' ? 'hot' : 'jsonb'}`,
+  schema: prototype === 'typed_eav'
+    ? 'idx_bench_eav'
+    : `idx_bench_${prototype === 'hot_projection' ? 'hot' : 'jsonb'}`,
   load_ms: 10 * scaleFactor,
   schema_bytes: 1_000 * scaleFactor,
   workloads: readWorkloads.map((name, index) => ({
@@ -98,6 +108,11 @@ const ratios = {
 const validComparison = () => ({
   generated_at: '2026-07-24T12:00:00Z',
   methodology: {
+    source_oracle: 'normalized idx_bench_source workload result digests',
+    result_digest: 'ordered_length_prefixed_json_v1',
+    evidence_validation: 'fail closed on report shape, metrics, plans, effects, ordering, digest semantics, and cardinalities',
+    first_run: 'first EXPLAIN ANALYZE repetition',
+    warm_run: 'median after the first repetition; not a guaranteed OS cold-cache comparison',
     automatic_winner_selection: false,
     comparable_database_fields: [...comparableDatabaseFields],
     database_settings_source: databaseSettingsSource,
@@ -142,20 +157,25 @@ const withFixture = (callback) => {
   }
 };
 
-const run = (root, comparison, decision) => {
+const invoke = (args) => spawnSync(process.execPath, [script, ...args], { encoding: 'utf8' });
+
+const run = (root, comparison, decision, extraArgs = []) => {
   const comparisonPath = path.join(root, 'comparison.json');
   const decisionPath = path.join(root, 'decision.json');
   const outputPath = path.join(root, 'adr.md');
   writeJson(comparisonPath, comparison);
   writeJson(decisionPath, decision);
-  const result = spawnSync('node', [
-    script,
+  const result = invoke([
     '--comparison', comparisonPath,
     '--decision', decisionPath,
     '--output', outputPath,
-  ], { encoding: 'utf8' });
-  return { result, outputPath };
+    ...extraArgs,
+  ]);
+  return { result, comparisonPath, decisionPath, outputPath };
 };
+
+const stagingEntries = (root) => readdirSync(root)
+  .filter((entry) => entry.startsWith('.adr.md.tmp-'));
 
 test('renders a manual same-commit storage ADR', () => {
   withFixture((root) => {
@@ -168,19 +188,21 @@ test('renders a manual same-commit storage ADR', () => {
     assert.match(markdown, /Comparison SHA-256/u);
     assert.match(markdown, /## Rejected alternatives/u);
     assert.match(markdown, /renderer does not infer or rank a winning prototype/u);
+    assert.deepEqual(stagingEntries(root), []);
   });
 });
 
 test('rejects core-only comparison without observed database-settings methodology', () => {
   withFixture((root) => {
     const comparison = validComparison();
-    comparison.methodology = { automatic_winner_selection: false };
-    const { result } = run(root, comparison, validDecision(comparison));
+    comparison.methodology.comparable_database_fields = ['server_version_num'];
+    const { result, outputPath } = run(root, comparison, validDecision(comparison));
     assert.notEqual(result.status, 0);
     assert.match(
       result.stderr,
       /comparable_database_fields must exactly match the canonical PostgreSQL database-settings contract/u,
     );
+    assert.equal(existsSync(outputPath), false);
   });
 });
 
@@ -188,9 +210,10 @@ test('rejects evidence that is not decision-ready', () => {
   withFixture((root) => {
     const comparison = validComparison();
     comparison.decision_ready = false;
-    const { result } = run(root, comparison, validDecision());
+    const { result, outputPath } = run(root, comparison, validDecision());
     assert.notEqual(result.status, 0);
     assert.match(result.stderr, /comparison is not decision-ready/u);
+    assert.equal(existsSync(outputPath), false);
   });
 });
 
@@ -235,5 +258,106 @@ test('rejects an unsatisfied evidence decision flag', () => {
     const { result } = run(root, comparison, decision);
     assert.notEqual(result.status, 0);
     assert.match(result.stderr, /same_result_digest_contract is not satisfied/u);
+  });
+});
+
+test('never overwrites the comparison input', () => {
+  withFixture((root) => {
+    const comparison = validComparison();
+    const decision = validDecision(comparison);
+    const comparisonPath = path.join(root, 'comparison.json');
+    const decisionPath = path.join(root, 'decision.json');
+    writeJson(comparisonPath, comparison);
+    writeJson(decisionPath, decision);
+    const original = readFileSync(comparisonPath);
+    const result = invoke([
+      '--comparison', comparisonPath,
+      '--decision', decisionPath,
+      '--output', comparisonPath,
+    ]);
+    assert.notEqual(result.status, 0);
+    assert.match(result.stderr, /--output must not overwrite the comparison input/u);
+    assert.deepEqual(readFileSync(comparisonPath), original);
+  });
+});
+
+test('never overwrites the decision input', () => {
+  withFixture((root) => {
+    const comparison = validComparison();
+    const decision = validDecision(comparison);
+    const comparisonPath = path.join(root, 'comparison.json');
+    const decisionPath = path.join(root, 'decision.json');
+    writeJson(comparisonPath, comparison);
+    writeJson(decisionPath, decision);
+    const original = readFileSync(decisionPath);
+    const result = invoke([
+      '--comparison', comparisonPath,
+      '--decision', decisionPath,
+      '--output', decisionPath,
+    ]);
+    assert.notEqual(result.status, 0);
+    assert.match(result.stderr, /--output must not overwrite the decision input/u);
+    assert.deepEqual(readFileSync(decisionPath), original);
+  });
+});
+
+test('help is successful only as the sole argument', () => {
+  const result = invoke(['--help']);
+  assert.equal(result.status, 0, result.stderr || result.stdout);
+  assert.match(result.stdout, /Usage:/u);
+});
+
+test('mixed help preserves an existing ADR output', () => {
+  withFixture((root) => {
+    const outputPath = path.join(root, 'adr.md');
+    const original = Buffer.from('reviewed ADR\n', 'utf8');
+    writeFileSync(outputPath, original);
+    const result = invoke(['--help', '--output', outputPath]);
+    assert.notEqual(result.status, 0);
+    assert.match(result.stderr, /help must be the only argument/u);
+    assert.deepEqual(readFileSync(outputPath), original);
+  });
+});
+
+test('unknown and duplicate options fail before changing output', () => {
+  withFixture((root) => {
+    const outputPath = path.join(root, 'adr.md');
+    const original = Buffer.from('reviewed ADR\n', 'utf8');
+    writeFileSync(outputPath, original);
+    const unknown = invoke([
+      '--comparison', 'missing-comparison.json',
+      '--decision', 'missing-decision.json',
+      '--output', outputPath,
+      '--format', 'markdown',
+    ]);
+    assert.notEqual(unknown.status, 0);
+    assert.match(unknown.stderr, /unknown or incomplete argument: --format/u);
+    assert.deepEqual(readFileSync(outputPath), original);
+
+    const duplicate = invoke([
+      '--comparison', 'missing-comparison.json',
+      '--decision', 'missing-decision.json',
+      '--output', outputPath,
+      '--output', path.join(root, 'other.md'),
+    ]);
+    assert.notEqual(duplicate.status, 0);
+    assert.match(duplicate.stderr, /--output was provided more than once/u);
+    assert.deepEqual(readFileSync(outputPath), original);
+  });
+});
+
+test('real render attempts revoke stale ADR output before core validation', () => {
+  withFixture((root) => {
+    const outputPath = path.join(root, 'adr.md');
+    writeFileSync(outputPath, 'stale ADR\n', 'utf8');
+    const result = invoke([
+      '--comparison', path.join(root, 'missing-comparison.json'),
+      '--decision', path.join(root, 'missing-decision.json'),
+      '--output', outputPath,
+    ]);
+    assert.notEqual(result.status, 0);
+    assert.match(result.stderr, /missing comparison/u);
+    assert.equal(existsSync(outputPath), false);
+    assert.deepEqual(stagingEntries(root), []);
   });
 });
